+++ src/renderer/viz/expressions/histogram/GlobalHistogram.js
@@ -1,5 +1,5 @@
 import Histogram from './Histogram';
-import { checkMaxArguments, implicitCast, checkArray } from '../utils';
+import { checkMaxArguments, implicitCast } from '../utils';
 import { CLUSTER_FEATURE_COUNT } from '../../../schema';
 export default class GlobalHistogram extends Histogram {
     constructor (property, sizeOrBuckets = 20, weight = 1) {
@@ -17,33 +17,20 @@ export default class GlobalHistogram extends Histogram {
             return this._cached;
         }
 
-        if (this.property.type === 'category') {
-
-        }
-
         this._cached = this.property.type === 'number'
             ? (this._hasBuckets ? this._getBucketsValue(this._histogram, this._sizeOrBuckets) : this._getNumericValue(this._histogram, this._sizeOrBuckets))
-            : this._getCategoryValue(this._histogram);
+            : this._getCategoryValue();
 
         return this._cached;
     }
 
-    getJoinedValues (values) {
-        checkArray('histogram.getJoinedValues', 'values', 0, values);
-
-        if (!values.length) {
-            return [];
-        }
-
-        return this.value.map(elem => {
-            const data = values.find(value => value.key === elem.x);
-
-            const frequency = elem.y;
-            const key = elem.x;
-            const value = data !== -1 ? data.value : null;
-
-            return { frequency, key, value };
+    _getCategoryValue () {
+        const categories = this._metadata.properties[this.property.name].categories;
+        categories.forEach(category => {
+            this._histogram.set(category.name, category.frequency);
         });
+
+        return super._getCategoryValue(this._histogram);
     }
 
     accumGlobalAgg (feature) {
@@ -54,15 +41,6 @@ export default class GlobalHistogram extends Histogram {
             const weight = clusterCount * this.weight.eval(feature);
             this._histogram.set(property, weight);
         }
-    }
-
-    _getCategoryValue () {
-        const categories = this._metadata.properties[this.property.name].categories;
-        categories.forEach(category => {
-            this._histogram.set(category.name, category.frequency);
-        });
-
-        return super._getCategoryValue(this._histogram);
     }
 
     _resetGlobalAgg (metadata) {
